Remove _BaseDatasetWrapperIterator
